cros-rust: avoid linking binary-only crates
If we're installing a binary-only create (e.g. crosh), then there
won't be any library creates to register.  Skip creating the symlink
as that will generate a dangling symlink that later breaks image test.

BUG=chromium:999931
TEST=`emerge-betty crosh` no longer installs symlink
TEST=`emerge-betty dev-rust/sync` still installs symlink
TEST=test_image on base image passes

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/1857346
Tested-by: Mike Frysinger <vapier@chromium.org>

+++ eclass/cros-rust.eclass
@@ -443,11 +443,15 @@ cros-rust_pkg_postinst() {
 
 	local crate_dir="${ROOT}${CROS_RUST_REGISTRY_DIR}/${crate}"
 	local registry_dir="${ROOT}${CROS_RUST_REGISTRY_INST_DIR}"
-	mkdir -p "${registry_dir}"
 
-	local dest="${registry_dir}/${crate}"
-	einfo "Linking ${crate} into Cargo registry at ${registry_dir}"
-	ln -srT "${crate_dir}" "${dest}" || die
+	# Don't install links for binary-only crates as they won't have any
+	# library crates to register.  This avoids dangling symlinks.
+	if [[ -e "${crate_dir}" ]]; then
+		local dest="${registry_dir}/${crate}"
+		einfo "Linking ${crate} into Cargo registry at ${registry_dir}"
+		mkdir -p "${registry_dir}"
+		ln -srT "${crate_dir}" "${dest}" || die
+	fi
 }
 
 # @FUNCTION: cros-rust_pkg_prerm
@@ -464,8 +468,12 @@ cros-rust_pkg_prerm() {
 	local crate="${name}-${version}"
 
 	local registry_dir="${ROOT}${CROS_RUST_REGISTRY_INST_DIR}"
-	einfo "Removing ${crate} from Cargo registry"
-	rm -f "${registry_dir}/${crate}" || die
+	local link="${registry_dir}/${crate}"
+	# Add a check to avoid spamming when it doesn't exist (e.g. binary crates).
+	if [[ -L "${link}" ]]; then
+		einfo "Removing ${crate} from Cargo registry"
+		rm -f "${link}" || die
+	fi
 }
 
 # @FUNCTION: cros-rust_get_crate_version
